Move providers to provider directory
