+++ app/login/page.tsx
@@ -3,7 +3,7 @@
 import { Auth } from '@supabase/auth-ui-react';
 import { useRouter } from 'next/navigation';
 
-import { useSupabase } from '@/app/supabase-provider';
+import { useSupabase } from '@/providers/supabase-provider';
 
 export default function Page() {
   const router = useRouter();
